Repo: Created components to repo in order to keep the code clean and slim.
