+++ src/components/Repos/index.js
@@ -1,0 +1,4 @@
+export { FiltersRepo } from './Filters';
+export { HeadersRepo } from './Header';
+export { IssuesRepo } from './Issues';
+export { FootersRepo } from './Footer';
